Move anomaly and change detector base classes to respective modules
